pkg/verify: Twiddle with the Store interface
And move the store interface and implementations down into
sub-packages.  Also drop the old locations business; pulling
signatures from the sig-store URI scheme is just another kind of
signature store.

This also adds name validation, to ensure that we only accept signed
OpenShift releases with the expected version, and not, say, a 4.1.0
image that a compromised Cincinnati is claiming is 4.3.0.  Or worse
yet, a completely unrelated image like the RHEL image from the verify
testdata.