WIP: install: Serve metrics over HTTPS
WIP because we still need to write Go backing --serving-cert-file and
--serving-key-file.

CVO metrics are currently being sent insecurely over HTTP.  Serve them
over HTTPS instead [1].

Add the serving-cert-secret-name annotation [1] to provision an X.509
keypair and cert so we can serve metrics over TLS/HTTPS.  This will
result in a cert that cluster components trust for
cluster-version-operator.openshift-cluster-version.svc being stored in
a secret named cluster-version-operator-serving-cert.

We then volume-mount that certificate into the production CVO pod (but
not into bootstrap/bootstrap-pod.yaml, because we want that to work
regardless of whether the annotation-consuming, secret-creating
handler is running).

bearerTokenFile and tlsConfig are defined (ish) in [3,4].  I'm not
really clear on how we are supposed to figure out these values, so
we're just blindly pattern-matching [5,6].

Recycling the same port may lead to some update-time hiccups, as the
new CVO is likely to come out serving HTTPS before the ServiceMonitor
is updated to declare HTTPS (and vice versa for downgrades).  Having
the ServiceMonitor at the end of the upgrade graph (runlevel 90) and
the CVO at the beginning means this drift could be extensive,
especially for updates that hang up.  We may want to claim a new port
for HTTPS and have the new CVO serve HTTP on 9099 and HTTPS on a
different port.  But 9099 is enshrined in the docs [7], and I dunno
what else is available.  I'm also not clear on if or why this would be
a node-level port.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1809195
[2]: https://github.com/openshift/openshift-docs/blob/07627a5f180f7823260bfa6f21fe467b1daa3903/modules/customize-certificates-add-service-serving.adoc
[3]: https://github.com/coreos/prometheus-operator/blob/0c6d184c19c890af10d715310d1770424d4e370b/Documentation/api.md#apiserverconfig
[4]: https://github.com/coreos/prometheus-operator/blob/0c6d184c19c890af10d715310d1770424d4e370b/Documentation/api.md#tlsconfig
[5]: https://github.com/openshift/cluster-openshift-apiserver-operator/commit/8a33e6931639d9f3afe50dbdc5d57c24a62577ad#diff-ec1060c1660b88fd2d140ee04ff421d9R8
[6]: https://github.com/openshift/cluster-openshift-apiserver-operator/commit/4e0470a19e25f02ae244918117130d4c3fea13c0
[7]: https://github.com/openshift/openshift-docs/blame/07627a5f180f7823260bfa6f21fe467b1daa3903/modules/installation-network-user-infra.adoc#L61