pkg: Add a KubeletVersionSkew update checker
Upstream docs [1]:

  kubelet must not be newer than kube-apiserver, and may be up to two
  minor versions older.

The coming OpenShift 4.8 is aiming at 1.21 [2].  This commit will end
up in 4.8.  So when it is considering Upgradeable guards, it is saying
"is the current state compatible with the next OpenShift minor?".  The
next minor will be 4.9, which will presumably run 1.22.  So it will be
compatible with 1.20, 1.21, and 1.22 kubelets.  If for some reason
OpenShift 4.9 ends up being compatible with 1.19 too (say, because it
sticks with 4.8's 1.21), we'll backport a patch to the 4.8
cluster-version operator to bring that wisdom back.  Those
next-minor-knowledge backports are the best we can do unless we get
something like preflight checks [3].

I'm not wildly excited about the current framework.  It has lots of
informers, and I think we might want a single NewSharedInformer to
reduce the number of parallel watches.  It also periodically polls all
of the upgradeable checks, and I would prefer a framework where our
informers fed queues and we had workers processing change events from
those queues to trigger state changes, which should give us lower
latency reactions.  But since we may end up taking this code back to
4.6, the current commit is a minimal addition following the existing
patterns, and we can consider more extensive refactors once we have
the backport-friendly pivot in place.

[1]: https://kubernetes.io/docs/setup/release/version-skew-policy/#kubelet
[2]: https://github.com/openshift/kubernetes/pull/641
[3]: https://github.com/openshift/enhancements/pull/363